pyproject.toml
[project]
name = "finsynapse"
version = "0.1.0"
description = "Personal multi-market macro data pipeline + market thermometer + local LLM insights"
authors = [
    { name = "hg.dendi", email = "hg.dendi@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "akshare>=1.18.59",
    "beautifulsoup4>=4.14.3",
    "empyrical-reloaded>=0.5.12",
    "jinja2>=3.1.6",
    "lxml>=6.1.0",
    "markdown>=3.10.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "pyarrow>=24.0.0",
    "pydantic-settings>=2.14.0",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "streamlit>=1.57.0",
    "typer>=0.25.0",
    "yfinance>=1.3.0",
]

[project.scripts]
finsynapse = "finsynapse.cli:app"

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "diff-cover>=10.2.0",
    "hypothesis>=6.152.4",
    "jsonschema>=4.26.0",
    "mutmut>=3.5.0",
    "pytest>=9.0.3",
    "pytest-benchmark>=5.2.3",
    "pytest-cov>=7.1.0",
    "pytest-vcr>=1.0.2",
    "ruff>=0.15.12",
    "scipy>=1.17.1",
    "syrupy>=5.1.0",
    "vcrpy>=8.1.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[tool.ruff]
# 120 chars matches the natural width of this codebase (longest existing
# lines are ~165 in CLI/i18n tables — those are intentional alignment).
# Anything wider gets the noqa hatch on a case-by-case basis.
line-length = 120
target-version = "py311"
src = ["src"]
extend-exclude = ["data", "dist", "scripts"]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes (unused imports, undefined names)
  "I",   # isort
  "W",   # pycodestyle warnings
  "UP",  # pyupgrade (modern syntax for our py311 floor)
  "B",   # flake8-bugbear (common bug patterns)
  "SIM", # flake8-simplify (simplify code)
  "RUF", # ruff-specific checks
]
ignore = [
  "E501",   # line-length is enforced by the formatter, not the linter
  "RUF001", # ambiguous fullwidth chars are intentional (Chinese i18n strings, math)
  "RUF002", # ambiguous dashes in docstrings are intentional (math/doc ranges)
  "RUF003", # ambiguous chars in comments are intentional (Chinese, Greek, math)
]

[tool.ruff.lint.per-file-ignores]
# i18n tables and CLI Typer signatures are intentionally wide for visual
# alignment; allow E741/E402 noise we don't intend to fix everywhere.
"src/finsynapse/dashboard/i18n.py" = ["E741"]
"tests/*" = ["F401", "F811"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
